// app/models/enums/enum-lib-group.ts
/*
 * 01. 구분     : Enum
 * 02. 타입     : -
 * 03. 업무구분  : 모든권한 - Enum
 * 03. 설명     : 다이얼로그 버튼 enum 제공
 * 04. 작성일자  : 2023.12.20
 * 05. 작성자   : 이희준
 */

export enum EnumDialogBtns {
  BTN_CANCEL_ID = 'BTN_CANCEL',
  BTN_CONFIRM_ID = 'BTN_CONFIRM',
  BTN_CANCEL_NM = '취소',
  BTN_CONFIRM_NM = '확인',
}